feat: better model for receipts

+++ src/main/java/com/manager/payments/model/payments/PaymentFactory.java
@@ -3,6 +3,7 @@ package com.manager.payments.model.payments;
 import com.manager.payments.model.exceptions.PaymentAlreadyExpired;
 import com.manager.payments.model.exceptions.PaymentInvalidDateInterval;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class PaymentFactory {
@@ -10,7 +11,7 @@ public class PaymentFactory {
     private PaymentFactory() {
     }
 
-    public static Payment build(String code, double amount, String name, String description, LocalDate startDate,
+    public static Payment build(String code, BigDecimal amount, String name, String description, LocalDate startDate,
                                 LocalDate endDate, Periodicity periodicity) {
         LocalDate now = LocalDate.now();
         if (startDate.isAfter(endDate)) {
